[Keyboard] add oled and encoder function to lck75 (#11825)
