+++ keyboards/lck75/rules.mk
@@ -29,3 +29,4 @@ ENCODER_ENABLE = yes
 WPM_ENABLE = yes
 LTO_ENABLE = no
 AUTO_SHIFT_ENABLE = no
+
